updatedb: introduce new option -d
If this option is specified on the command line then /proc/filesystems file is
parsed and all filesystems which are marked as "nodev" are added to file of
filesystem types we prune.

There are two exceptions though, zfs and rootfs. Those are marked as nodev but we
really do want to index them. Thus we will not add them to prune list.

AFAIK, file /proc/filesystems is Linux specific API. In case we can't open the
file we will just skip the whole procedure but updatedb *will not* report any error.